Factor scores and IRT estimates of latent dimensions using the gwf data.

The factor scores are replicated from Wright, Joseph. 2021. “The Latent Characteristics That Structure Autocratic Rule.” Political Science Research and Methods 9 (1): 1–19. https://doi.org/10.1017/psrm.2019.50. See the vignette for details.

Usage

gwf_irt

Format

An object of class tbl_df (inherits from tbl, data.frame) with 4591 rows and 19 columns.

Details

gwf_country

name of country

cowcode

Correlates of War country code. Countries that have changed territory during the period from 1946 to 2010 have the following country codes in this data set: Czechoslovakia (315); East Germany (265); Ethiopia (530); Eritrea (531); Russia and Soviet Union (365); South Vietnam (817); South Yemen (680); Vietnam/North Vietnam (816); Yemen/ North Yemen (678); Yugoslavia/Serbia (435). Note that regimes that either lost (Ethiopia) or gained territory (Vietnam/North Vietnam) during the lifetime of the regime have the same country code throughout the entire regime spell. Users should note this feature of the data set when merging with other data sets.

year

calendar year

pers_irt_11_F1

IRT scores for personalism, 11 items

pers_irt_11_F1

SE for IRT scores for personalism, 11 items

pers_irt_10_F1

IRT scores for personalism, 10 items

pers_irt_10_F1

SE for IRT scores for personalism, 10 items

pers_irt_8_F1

IRT scores for personalism, 8 items

pers_irt_8_F1

SE for IRT scores for personalism, 8 items

pers_irt_11_full_info_F1

IRT scores for personalism, full information, 11 items

pers_irt_11_full_info_F1

SE for IRT scores for personalism, full information, 11 items

party_irt_11_F1

IRT scores for party, 11 items

party_irt_11_F1

SE for IRT scores for party, 11 items

military_irt_11_F1

IRT scores for party, 11 items

military_irt_11_F1

SE for IRT scores for party, 11 items

party_factor

Scores on first factor of factor analysis (party).

military_factor

Scores on second factor of factor analysis (military).

personalism_factor

Scores on third factor of factor analysis (personalism).

fourth_factor

Scores on fourth factor of factor analysis (other).